// src/models/Car.ts
import dayjs, { Dayjs } from "dayjs";
export interface Car {
  id: number;
  parkingLotId: number;
  carStatusId: number | null;
  carStatus: string | null;
  carId: number | null;
  carLicensePlates: string;
  seatNumber: number;
  carMakeId: string;
  carModelId: string;
  modelYear: number | null;
  carGenerationId: string;
  carSeriesId: string;
  carTrimId: string;
  makeName: string | null;
  modelName: string | null;
  generationName: string | null;
  seriesName: string | null;
  trimName: string | null;
  generationYearBegin: number | null;
  generationYearEnd: number | null;
  trimStartProductYear: null;
  trimEndProductYear: null;
  createdDate: Date | null;
  isDeleted: boolean;
  carColor: string;
  carFuel: string;
  priceForNormalDay: number;
  priceForWeekendDay: number;
  priceForMonth: number;
  limitedKmForMonth: number;
  overLimitedMileage: number;
  carStatusDescription: string | null;
  currentEtcAmount: number;
  carDescription: string;
  fuelPercent: number;
  speedometerNumber: number;
  carOwnerName: string;
  rentalMethod: string;
  rentalDate: Date | Dayjs | null | undefined;
  speedometerNumberReceive: number;
  priceForDayReceive: number;
  // priceForDayReceive: string;
  priceForMonthReceive: number;
  insurance: boolean;
  limitedKmForMonthReceive: number;
  overLimitedMileageReceive: number;
  frontImg: string;
  backImg: string;
  leftImg: string;
  rightImg: string;
  ortherImg: string;
  carFileCreatedDate: Date | null;
  linkTracking: string;
  trackingUsername: string;
  trackingPassword: string;
  etcusername: string;
  etcpassword: string;
  linkForControl: null;
  paymentMethod: null;
  forControlDay: null;
  dayOfPayment: null;
  periodicMaintenanceLimit: number;
  carSchedules: [] | null;
  parkingLotName: string | null;
  registrationDeadline: Date | Dayjs | null | undefined;
  carKmLastMaintenance: number | null;
  ownerSlitRatio: number;
  kmTraveled: number | null;
  maintenanceDate: Date | Dayjs | null | undefined;
  maintenanceInvoice: string;
  maintenanceAmount: number | null;
  carMakeName: string | null;
  carModelName: string | null;
  carGenerationName: string | null;
  carSeriesName: string | null;
  carTrimName: string | null;
   tankCapacity:number ;
}